+++ juriscraper/opinions/united_states/state/alaska_slip.py
@@ -1,6 +1,7 @@
 from datetime import datetime
 
 from juriscraper.opinions.united_states.state import alaska
+from juriscraper.OpinionSite import OpinionSite
 
 
 class Site(alaska.Site):
@@ -9,6 +10,7 @@ class Site(alaska.Site):
     is_coa = False
     search_parameter = "PublishedOrders"
     document_type = "PUBORDER"
+    extract_from_text = OpinionSite.extract_from_text
 
     def __init__(self, *args, **kwargs):
         super().__init__(*args, **kwargs)