+++ CHANGES.md
@@ -21,9 +21,10 @@ Features:
 Changes:
 - Retrieve lower court information in `sd` #1569
 - Retrieve lower court information in `va` #1569
+- Retrieve lower court information in `utah` #1569
 
 Fixes:
--
+- fix `bia` scraper #1574
 
 ## Current