+++ CHANGES.md
@@ -20,6 +20,7 @@ Features:
 
 Changes:
 - Retrieve lower court information in `sd` #1569
+- Retrieve lower court information in `va` #1569
 - Retrieve lower court information in `utah` #1569
 
 Fixes: