+++ CHANGES.md
@@ -24,6 +24,7 @@ Changes:
 - Retrieve lower court information in `utah` #1569
 
 Fixes:
+- Fix connappct #1580
 - fix `bia` scraper #1574
 
 ## Current